Move Schema Manager to get rid of cyclic imports
